Use a dummy PWM timer to generate the pwm-period / tick interrupt.
This way, we can use a true 0% duty cycle for the real PWM timers.

I did not find a good way to trigger the interrupt from a timer that
is running at 0% duty cycle - seems the timeout interrupt does not
trigger in PWM mode, and with 0% duty cycle there is no rising or
falling edge on which to trigger the event interrupt.