Start ADC measurements early in the PWM cycle.
This helps ensure that the ADC measurement is made during the active part of
the duty cycle, where one phase is still high.

For reasons that I still don't fully understand, the back-EMF measurements
only make sense if made while one of the other phases is high and one is
low.