Improved sampling of back-EMF.
Implement more detailed debugging. We sample the back-EMF 8 times in a row
(at 1 MHz sample rate), to see how the signal varies over a single PWM
period. Results show that it takes a few microseconds for the back-EMF
signal to stabilise at the start of a PWM period, so it seems preferential
to measure towards the end of the duty cycle.

Reduce commutation frequency to 25 kHz (this doubles the on-time for a given
duty cycle, giving more time for back-EMF signal to stabilise). Use a
default duty cycle of 20%, which seems better to allow back-EMF to
stabilise.