Implement controlled spin-down of motor.
A switch (on the PS2 controller board) controls spin-up and spin-down.

Spin-up is the same open-loop that goes to closed-loop once a certain speed
is reached.

Spin-down works by lovering the voltage (motor_damper) gradually to zero and
then waiting for 10 seconds for the motor to stop on its own.

One idea is that gentle spin-down should avoid nasty voltage spikes when the
coils are suddenly cut-off, in case that might kill the MCU (won't help
against power outage though, probably).

This commit also re-factors the motor control code a bit to make things
somewhat clearer as to what goes on in each stage of the motor control. It
is still not perfect, though.