Use measured voltage divider bias to correct back-emf debug output
To do this, we store now measured phase and neutral separately in the debug
buffer, along with a flag to say which phase was stored.

Then for debug printout, we can use the measured correction factor to get
the measured voltage, before subtracting to get the back-emf value.

(This commit only changes the debug output; the zero-crossing detection
still uses values not corrected for bias).