Improved zero-crossing detection for deciding next commute step
Use back-emf measurements corrected for measured bias, when detecting
zero-crossings.

And commute two pwm periods earlier, this seems to give more symmetric
commutation relative to the back-emf. I guess there is a couple of steps
delay in the measurement of back-emf and the setup of next commute step, so
two steps worth of correction seems reasonable.